+++ src/v2/Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarCommercial.tsx
@@ -1,4 +1,4 @@
-import { ContextModule, Intent, OwnerType } from "@artsy/cohesion"
+import { ContextModule, Intent } from "@artsy/cohesion"
 import {
   Box,
   Button,
@@ -31,17 +31,11 @@ import {
 import { ErrorWithMetadata } from "v2/Utils/errors"
 import { get } from "v2/Utils/get"
 import createLogger from "v2/Utils/logger"
-import { AuthModalOptions, openAuthModal } from "v2/Utils/openAuthModal"
+import { openAuthModal } from "v2/Utils/openAuthModal"
 import { ArtworkSidebarSizeInfoFragmentContainer as SizeInfo } from "./ArtworkSidebarSizeInfo"
 import { Mediator } from "lib/mediator"
 import { useInquiry, WithInquiryProps } from "v2/Components/Inquiry/useInquiry"
-import { SavedSearchCreateAlertBase } from "v2/Components/SavedSearchAlert/Components/SavedSearchCreateAlertBase"
-import {
-  SavedSearchEntity,
-  SearchCriteriaAttributes,
-} from "v2/Components/SavedSearchAlert/types"
-import { compact } from "lodash"
-import { getAttributionClassValueByLabel } from "v2/Components/SavedSearchAlert/Utils/getAttributionClassValueByLabel"
+import { ArtworkSidebarCreateAlertButtonFragmentContainer } from "./ArtworkSidebarCreateAlertButton"
 
 type EditionSet = NonNullable<
   ArtworkSidebarCommercial_artwork["edition_sets"]
@@ -373,26 +367,6 @@ export class ArtworkSidebarCommercialContainer extends React.Component<
     }
   }
 
-  getAuthModalOptions(): AuthModalOptions {
-    const { artwork } = this.props
-
-    return {
-      entity: {
-        name: artwork.title!,
-        slug: artwork.slug,
-      },
-      afterSignUpAction: {
-        action: "createAlert",
-        kind: "artworks",
-        objectId: artwork.internalID,
-      },
-      contextModule: ContextModule.artworkSidebar,
-      intent: Intent.createAlert,
-      redirectTo: location.href,
-      copy: "Sign up to create your alert",
-    }
-  }
-
   render() {
     const { artwork, inquiryComponent } = this.props
     const {
@@ -423,31 +397,6 @@ export class ArtworkSidebarCommercialContainer extends React.Component<
 
     const shouldDisplayMakeOfferAsPrimary: boolean | null =
       shouldDisplayMakeOfferButton && isInquireable
-
-    const attributionClass = getAttributionClassValueByLabel(
-      artwork.attributionClass?.name ?? ""
-    )
-    const artists = compact(artwork.artists)
-    const artistIDs = artists.map(artist => artist.internalID)
-    const placeholder = `Artworks like: ${artwork.title!}`
-    const entity: SavedSearchEntity = {
-      placeholder,
-      artists: artists.map(artist => ({
-        id: artist.internalID,
-        name: artist.name ?? "",
-        slug: artist.slug,
-      })),
-      owner: {
-        type: OwnerType.artwork,
-        slug: artwork.slug,
-        id: artwork.internalID,
-        name: artwork.title!,
-      },
-    }
-    const criteria: SearchCriteriaAttributes = {
-      artistIDs,
-      attributionClass,
-    }
 
     return (
       <>
@@ -540,16 +489,7 @@ export class ArtworkSidebarCommercialContainer extends React.Component<
             </>
           )}
 
-          <SavedSearchCreateAlertBase
-            entity={entity}
-            criteria={criteria}
-            getAuthModalOptions={this.getAuthModalOptions}
-            renderButton={({ onClick }) => (
-              <Button width="100%" size="medium" onClick={onClick}>
-                Create Alert
-              </Button>
-            )}
-          />
+          <ArtworkSidebarCreateAlertButtonFragmentContainer artwork={artwork} />
 
           {isInquireable && (
             <>
@@ -640,15 +580,7 @@ export const ArtworkSidebarCommercialFragmentContainer = createFragmentContainer
         shippingInfo
         shippingOrigin
         slug
-        title
-        artists {
-          internalID
-          name
-          slug
-        }
-        attributionClass {
-          name
-        }
+        ...ArtworkSidebarCreateAlertButton_artwork
       }
     `,
   }
